Extract TokenUsage component
